CAFFIENE_CLAUSE__SENTINAL_GEC/requirements.txt
django
djangorestframework
django-cors-headers
python-decouple
web3
opencv-python-headless
ultralytics
requests
